Feat: 총 재고수량 상태관리 추가, 머지된 컴포넌트 style 수정 - 겹치는 보더 삭제 - 중복 마진 값 삭제 - 병합된 molecules 단위 컴포넌트 css 조정
Feat: 총 재고수량 전역 상탯값 추가

Chore: 출력문 삭제
