+++ src/components/molecules/BasicInfo/MainImg.jsx
@@ -1,18 +1,20 @@
 import ImageUpload from 'components/atoms/ImageUpload';
 import SettingFramItem from 'components/atoms/SettingFrameItem';
 import React, { useEffect, useState } from 'react';
+import styled from 'styled-components';
 import PropTypes from 'prop-types';
 
 function MainImg({ handleBasicInfo, title }) {
   const [mainImages, setMainImages] = useState([]);
   useEffect(() => {
-    // console.log(title, mainImages);
     handleBasicInfo(title, mainImages);
   }, [mainImages]);
 
   return (
     <SettingFramItem title="상품 대표 이미지">
-      <ImageUpload imageFiles={mainImages} setImageFiles={setMainImages} type="multiple" />
+      <Wrapper>
+        <ImageUpload imageFiles={mainImages} setImageFiles={setMainImages} type="multiple" />
+      </Wrapper>
     </SettingFramItem>
   );
 }
@@ -21,5 +23,9 @@ MainImg.propTypes = {
   title: PropTypes.string.isRequired,
   handleBasicInfo: PropTypes.func.isRequired,
 };
+
+const Wrapper = styled.div`
+  height: 10rem;
+`;
 
 export default MainImg;
