+++ src/components/molecules/BasicInfo/ProductCode.jsx
@@ -1,5 +1,4 @@
 import SettingFramItem from 'components/atoms/SettingFrameItem';
-import { BORDER_COLOR } from 'constants/color';
 import React, { useEffect } from 'react';
 import styled from 'styled-components';
 import PropTypes from 'prop-types';
@@ -10,13 +9,11 @@ function ProductCode({ handleBasicInfo, title }) {
     handleBasicInfo(title, code);
   }, []);
   return (
-    <Wrapper>
-      <SettingFramItem title="상품코드">
-        <Container>
-          <Code>{code}</Code>
-        </Container>
-      </SettingFramItem>
-    </Wrapper>
+    <SettingFramItem title="상품코드">
+      <Wrapper>
+        <p>{code}</p>
+      </Wrapper>
+    </SettingFramItem>
   );
 }
 
@@ -26,20 +23,10 @@ ProductCode.propTypes = {
 };
 
 const Wrapper = styled.div`
-  height: 100%;
-  border-left: 1px solid ${BORDER_COLOR};
-`;
-
-const Container = styled.div`
   display: flex;
   align-items: center;
   padding: 12.8px 8px 12.8px 8px;
-  height: 115%;
-  border-left: 1px solid ${BORDER_COLOR};
-`;
-
-const Code = styled.div`
-  height: 100%;
+  height: 4.1rem;
 `;
 
 export default ProductCode;
