+++ src/components/molecules/BasicInfo/TotalStock.jsx
@@ -1,12 +1,18 @@
 import SettingFramItem from 'components/atoms/SettingFrameItem';
 import React from 'react';
+import { useRecoilValue } from 'recoil';
 import styled from 'styled-components';
+import { totalStockState } from 'utils/globalState';
 
 function TotalStock() {
+  const totalStock = useRecoilValue(totalStockState);
   return (
     <SettingFramItem title="상품 총 재고">
       <StockContainer>
-        <p>dd개</p>
+        <p>
+          {totalStock}
+          개
+        </p>
       </StockContainer>
     </SettingFramItem>
   );
